# docs/api/openapi/cordum-rest.yaml
openapi: 3.0.3
info:
  title: Cordum REST API
  version: 1.0.0
  description: |
    Safety-first agent orchestration platform. Cordum provides a control plane
    for submitting, scheduling, and governing AI agent jobs with built-in safety
    policies, approval workflows, dead-letter queues, and audit logging.

    ## Authentication
    All protected endpoints require either an API key (`X-API-Key` header) or a
    Bearer JWT token obtained from the login endpoint. Public endpoints are
    explicitly marked and require no credentials.

    ## Multi-Tenancy
    Protected routes require the `X-Tenant-ID` header to scope requests to the
    correct tenant partition.
  contact:
    name: Cordum Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8081
    description: Development
...
security:
  - apiKey: []
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/auth/config:
    get:
      operationId: getAuthConfig
      summary: Get authentication configuration
      description: Returns the authentication methods enabled for the deployment.
      tags: [Auth]
      security: []
      responses:
        "200":
          description: Auth configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthConfig"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/config:
    get:
      operationId: getConfig
      summary: Get configuration
      tags: [Config]
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - name: scope
          in: query
          schema:
            type: string
            enum: [global, org, team, workflow, step]
        - name: scope_id
          in: query
          schema:
            type: string
        - name: envelope
          in: query
          description: Whether to wrap in {scope, data} envelope
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Configuration data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: setConfig
      summary: Set configuration
      tags: [Config]
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigDocument"
      responses:
        "204":
          description: Configuration saved
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued by the Cordum admin.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login.

  # =========================================================================
  # Parameters
  # =========================================================================
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant isolation header (required on all protected routes).
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized — missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden — insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
      required: [error, status]
...
    AuthConfig:
      type: object
      properties:
        password_enabled:
          type: boolean
        user_auth_enabled:
          type: boolean
        saml_enabled:
          type: boolean
        oidc_enabled:
          type: boolean
        require_2fa:
          type: boolean
        session_ttl_seconds:
          type: integer
...
    ConfigDocument:
      type: object
      properties:
        scope:
          type: string
          enum: [global, org, team, workflow, step]
        scope_id:
          type: string
        data:
          type: object
          additionalProperties: true
          description: Configuration key-value pairs

    # --- Schemas ----------------------------------------------------------
...